preparing to test isolating spack
