+++ pakages/__init__.py
@@ -1,6 +1,3 @@
 from pakages.version import __version__
 
 assert __version__
-import pakages.utils as utils
-
-utils.add_spack_to_path()
