+++ pakages/cli/build.py
@@ -2,20 +2,16 @@ __author__ = "Vanessa Sochat, Alec Scott"
 __copyright__ = "Copyright 2021-2022, Vanessa Sochat and Alec Scott"
 __license__ = "Apache-2.0"
 
-from pakages.client import PakClient
 from pakages.logger import logger
-from .helpers import parse_package_request
-
+from pakages.client import get_client
 
 def main(args, parser, extra, subparser):
-    cli = PakClient(args.settings_file)
+    cli = get_client(builder=args.builder, settings_file=args.settings_file)
 
     # Get list of packages and (optionally) repository
-    packages, repo = parse_package_request(args)
+    packages = cli.parse_package_request(args)
 
     # Finally, add the repository and install packages
-    if repo:
-        cli.add_repository(repo.repo_dir)
     cache = cli.build(
         args.packages,
         cache_dir=args.cache_dir,
