+++ pakages/cli/shell.py
@@ -2,8 +2,7 @@ __author__ = "Vanessa Sochat, Alec Scott"
 __copyright__ = "Copyright 2021-2022, Vanessa Sochat and Alec Scott"
 __license__ = "Apache-2.0"
 
-from pakages.client import PakClient
-
+from pakages.client import get_client
 
 def main(args, parser, extra, subparser):
 
@@ -29,8 +28,7 @@ def main(args, parser, extra, subparser):
 
 
 def create_client(args):
-    return PakClient(settings_file=args.settings_file)
-
+    return get_client(builder=args.builder, settings_file=args.settings_file)
 
 def ipython(args):
     """
