+++ pakages/spack/__init__.py
@@ -1,0 +1,2 @@
+from .utils import add_spack_to_path
+add_spack_to_path()
